requirements.txt
tensorflow
numpy
pillow
scikit-learn
matplotlib
fastapi
uvicorn